Align Canvas runtime ownership with core decomposition
