+++ scripts/core-boundaries/rules/feature-rules.mjs
@@ -55,7 +55,7 @@ export const optionalDependencyFeatureOwnerRules = [
         ownerFeatures: ['mcp', 'miniapp-runtime', 'remote-connect', 'remote-ssh-concrete'],
       },
       { depName: 'bitfun-agent-runtime', ownerFeatures: ['deep-research'] },
-      { depName: 'bitfun-product-domains', ownerFeatures: ['function-agents', 'miniapp-runtime'] },
+      { depName: 'bitfun-product-domains', ownerFeatures: ['canvas-runtime', 'function-agents', 'miniapp-runtime'] },
       { depName: 'bitfun-runtime-ports', ownerFeatures: ['remote-connect'] },
       {
         depName: 'bitfun-services-core',
@@ -73,6 +73,7 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'mac_address', ownerFeatures: ['remote-connect'] },
       { depName: 'md5', ownerFeatures: ['remote-connect'] },
       { depName: 'notify', ownerFeatures: ['file-watch'] },
+      { depName: 'oxc', ownerFeatures: ['canvas-runtime'] },
       { depName: 'qrcode', ownerFeatures: ['remote-connect'] },
       { depName: 'rand', ownerFeatures: ['mcp', 'remote-connect', 'remote-ssh-concrete'] },
       { depName: 'reqwest', ownerFeatures: ['announcement', 'browser-control', 'debug-log', 'mcp', 'miniapp-runtime', 'remote-connect', 'review-platform', 'web-tools'] },
@@ -83,7 +84,7 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'rustls', ownerFeatures: ['remote-connect'] },
       { depName: 'rustls-native-certs', ownerFeatures: ['remote-connect'] },
       { depName: 'schannel', ownerFeatures: ['remote-connect'] },
-      { depName: 'sha2', ownerFeatures: ['remote-connect', 'remote-ssh'] },
+      { depName: 'sha2', ownerFeatures: ['canvas-runtime', 'remote-connect', 'remote-ssh'] },
       { depName: 'shellexpand', ownerFeatures: ['remote-ssh-concrete'] },
       { depName: 'sse-stream', ownerFeatures: ['mcp'] },
       { depName: 'ssh_config', ownerFeatures: ['remote-ssh-concrete', 'ssh_config'] },
@@ -91,8 +92,8 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'thiserror', ownerFeatures: ['browser-control', 'git', 'remote-ssh-concrete', 'review-platform', 'web-tools', 'workspace-search'] },
       { depName: 'tokio-tungstenite', ownerFeatures: ['remote-connect'] },
       { depName: 'tokio-util', ownerFeatures: ['remote-ssh'] },
-      { depName: 'urlencoding', ownerFeatures: ['remote-connect'] },
-      { depName: 'uuid', ownerFeatures: ['miniapp-runtime', 'remote-connect', 'remote-ssh-concrete'] },
+      { depName: 'urlencoding', ownerFeatures: ['canvas-runtime', 'remote-connect'] },
+      { depName: 'uuid', ownerFeatures: ['canvas-runtime', 'miniapp-runtime', 'remote-connect', 'remote-ssh-concrete'] },
       { depName: 'which', ownerFeatures: ['miniapp-runtime', 'remote-connect', 'workspace-search'] },
       { depName: 'x25519-dalek', ownerFeatures: ['remote-connect'] },
     ],
@@ -161,6 +162,7 @@ export const ownerCrateFeatureAssemblyRules = [
       'computer-use',
       'image-analysis',
       'miniapp',
+      'canvas',
       'agent-control',
     ],
   },
@@ -170,6 +172,7 @@ export const ownerCrateFeatureAssemblyRules = [
     requiredProductFullFeatures: [
       'announcement',
       'browser-control',
+      'canvas-runtime',
       'debug-log',
       'deep-research',
       'file-watch',
