+++ src/crates/assembly/core/Cargo.toml
@@ -49,7 +49,6 @@ tower-http = { workspace = true, optional = true }
 
 glob = { workspace = true, optional = true }
 notify = { workspace = true }
-oxc = { workspace = true, optional = true }
 dirs = { workspace = true }
 dunce = { workspace = true }
 filetime = { workspace = true, optional = true }
@@ -157,7 +156,7 @@ schannel = "0.1"
 default = ["product-full"]
 product-full = [
     "ai-adapter-runtime",
-    "canvas-compiler",
+    "canvas-runtime",
     "dep:chrono-tz",
     "dep:cron",
     "dep:dashmap",
@@ -192,6 +191,10 @@ product-domains = [
     "bitfun-services-integrations/miniapp-runtime",
     "bitfun-product-domains/product-full",
 ]
+canvas-runtime = [
+    "product-domains",
+    "bitfun-services-integrations/canvas-runtime",
+]
 runtime-services = []
 service-integrations = [
     "dep:aes-gcm",
@@ -210,7 +213,6 @@ service-integrations = [
 ]
 tool-packs = ["dep:bitfun-tool-packs", "bitfun-tool-packs/product-full", "dep:image"]
 tauri-support = ["tauri"]  # Optional tauri support
-canvas-compiler = ["dep:oxc", "product-domains"]
 ssh-remote = [
     "bitfun-services-integrations/remote-ssh-concrete",
     "russh",
@@ -218,6 +220,22 @@ ssh-remote = [
 
 [dev-dependencies]
 tempfile = { workspace = true }
+
+[[test]]
+name = "context_profile"
+required-features = ["product-full"]
+
+[[test]]
+name = "git_contracts"
+required-features = ["service-integrations"]
+
+[[test]]
+name = "product_assembly"
+required-features = ["product-full"]
+
+[[test]]
+name = "remote_mcp_streamable_http"
+required-features = ["service-integrations"]
 
 [build-dependencies]
 sha2 = { workspace = true }
