+++ src/crates/assembly/core/src/agentic/agents/definitions/modes/cowork.rs
@@ -52,14 +52,6 @@ impl CoworkMode {
                 "WebFetch".to_string(),
                 "ControlHub".to_string(),
                 "InitMiniApp".to_string(),
-                #[cfg(feature = "product-domains")]
-                "CreateCanvas".to_string(),
-                #[cfg(feature = "product-domains")]
-                "ReadCanvas".to_string(),
-                #[cfg(feature = "product-domains")]
-                "UpdateCanvas".to_string(),
-                #[cfg(feature = "product-domains")]
-                "PatchCanvas".to_string(),
             ],
         }
     }
@@ -116,13 +108,5 @@ mod tests {
     fn cowork_mode_includes_init_miniapp_in_default_tools() {
         let tools = CoworkMode::new().default_tools();
         assert!(tools.contains(&"InitMiniApp".to_string()));
-        #[cfg(feature = "product-domains")]
-        assert!(tools.contains(&"CreateCanvas".to_string()));
-        #[cfg(feature = "product-domains")]
-        assert!(tools.contains(&"ReadCanvas".to_string()));
-        #[cfg(feature = "product-domains")]
-        assert!(tools.contains(&"UpdateCanvas".to_string()));
-        #[cfg(feature = "product-domains")]
-        assert!(tools.contains(&"PatchCanvas".to_string()));
     }
 }
