+++ src/crates/assembly/core/src/agentic/tools/implementations/mod.rs
@@ -3,7 +3,7 @@
 pub mod analyze_image_tool;
 pub mod ask_user_question_tool;
 pub mod bash_tool;
-#[cfg(feature = "product-domains")]
+#[cfg(feature = "canvas-runtime")]
 pub mod canvas_tools;
 pub mod code_review_tool;
 pub mod computer_use_actions;
@@ -50,7 +50,7 @@ pub use crate::agentic::tools::product_runtime::GetToolSpecTool;
 pub use analyze_image_tool::AnalyzeImageTool;
 pub use ask_user_question_tool::AskUserQuestionTool;
 pub use bash_tool::BashTool;
-#[cfg(feature = "product-domains")]
+#[cfg(feature = "canvas-runtime")]
 pub use canvas_tools::{CreateCanvasTool, PatchCanvasTool, ReadCanvasTool, UpdateCanvasTool};
 pub use code_review_tool::CodeReviewTool;
 pub use computer_use_tool::ComputerUseTool;
