+++ src/crates/assembly/core/src/agentic/tools/product_runtime/materialization.rs
@@ -35,13 +35,13 @@ impl StaticToolProviderFactory<dyn Tool> for ProductConcreteToolFactory {
             "get_goal" => Some(Arc::new(GetGoalTool::new())),
             "create_goal" => Some(Arc::new(CreateGoalTool::new())),
             "update_goal" => Some(Arc::new(UpdateGoalTool::new())),
-            #[cfg(feature = "product-domains")]
+            #[cfg(feature = "canvas-runtime")]
             "CreateCanvas" => Some(Arc::new(CreateCanvasTool::new())),
-            #[cfg(feature = "product-domains")]
+            #[cfg(feature = "canvas-runtime")]
             "ReadCanvas" => Some(Arc::new(ReadCanvasTool::new())),
-            #[cfg(feature = "product-domains")]
+            #[cfg(feature = "canvas-runtime")]
             "UpdateCanvas" => Some(Arc::new(UpdateCanvasTool::new())),
-            #[cfg(feature = "product-domains")]
+            #[cfg(feature = "canvas-runtime")]
             "PatchCanvas" => Some(Arc::new(PatchCanvasTool::new())),
             "CreatePlan" => Some(Arc::new(CreatePlanTool::new())),
             "submit_code_review" => Some(Arc::new(CodeReviewTool::new())),
