+++ src/crates/assembly/core/src/agentic/tools/registry.rs
@@ -356,15 +356,15 @@ mod tests {
             "get_goal",
             "create_goal",
             "update_goal",
-            "CreateCanvas",
-            "ReadCanvas",
-            "UpdateCanvas",
-            "PatchCanvas",
             "CreatePlan",
             "submit_code_review",
             "GetToolSpec",
             "GetFileDiff",
             "Log",
+            "CreateCanvas",
+            "ReadCanvas",
+            "UpdateCanvas",
+            "PatchCanvas",
             "SessionControl",
             "SessionMessage",
             "SessionHistory",
@@ -585,12 +585,12 @@ mod tests {
                 "AskUserQuestion",
                 "TodoWrite",
                 "get_goal",
-                "ReadCanvas",
                 "CreatePlan",
                 "submit_code_review",
                 "GetToolSpec",
                 "GetFileDiff",
                 "Log",
+                "ReadCanvas",
                 "SessionHistory",
                 "WebSearch",
                 "WebFetch",
