+++ src/crates/execution/tool-provider-groups/Cargo.toml
@@ -18,5 +18,6 @@ browser-web = []
 computer-use = []
 image-analysis = []
 miniapp = []
+canvas = []
 agent-control = []
-product-full = ["basic", "git", "mcp", "browser-web", "computer-use", "image-analysis", "miniapp", "agent-control"]
+product-full = ["basic", "git", "mcp", "browser-web", "computer-use", "image-analysis", "miniapp", "canvas", "agent-control"]
