+++ src/crates/services/services-integrations/Cargo.toml
@@ -36,6 +36,7 @@ local-ip-address = { workspace = true, optional = true }
 mac_address = { workspace = true, optional = true }
 md5 = { workspace = true, optional = true }
 notify = { workspace = true, optional = true }
+oxc = { workspace = true, optional = true }
 qrcode = { workspace = true, optional = true }
 rand = { workspace = true, optional = true }
 reqwest = { workspace = true, optional = true }
@@ -69,6 +70,13 @@ schannel = { version = "0.1", optional = true }
 default = []
 announcement = ["reqwest"]
 browser-control = ["anyhow", "bitfun-services-core", "dirs", "reqwest", "thiserror"]
+canvas-runtime = [
+    "dep:bitfun-product-domains",
+    "oxc",
+    "sha2",
+    "urlencoding",
+    "uuid",
+]
 debug-log = ["reqwest"]
 deep-research = ["bitfun-agent-runtime"]
 git = ["bitfun-services-core", "chrono", "git2", "thiserror"]
@@ -164,6 +172,7 @@ web-tools = ["reqwest", "thiserror"]
 product-full = [
     "announcement",
     "browser-control",
+    "canvas-runtime",
     "debug-log",
     "deep-research",
     "file-watch",
