+++ src/crates/services/services-integrations/src/canvas/compiler/runtime_bootstrap.js
