+++ src/crates/services/services-integrations/src/canvas/compiler/tests.rs
@@ -1,13 +1,13 @@
-#[cfg(feature = "canvas-compiler")]
+#[cfg(feature = "canvas-runtime")]
 use super::*;
-#[cfg(feature = "canvas-compiler")]
+#[cfg(feature = "canvas-runtime")]
 use bitfun_product_domains::canvas::types::{CanvasId, CanvasRevision};
 
-#[cfg(feature = "canvas-compiler")]
+#[cfg(feature = "canvas-runtime")]
 mod enabled {
     use super::*;
 
-    #[cfg(feature = "canvas-compiler")]
+    #[cfg(feature = "canvas-runtime")]
     fn source(source: &str) -> CanvasSource {
         CanvasSource::new_tsx(
             CanvasId::new("canvas_1"),
@@ -756,10 +756,10 @@ export default function Canvas() {
     }
 }
 
-#[cfg(not(feature = "canvas-compiler"))]
+#[cfg(not(feature = "canvas-runtime"))]
 use super::*;
 
-#[cfg(not(feature = "canvas-compiler"))]
+#[cfg(not(feature = "canvas-runtime"))]
 #[test]
 fn canvas_compiler_reports_feature_disabled_without_canvas_feature() {
     let diagnostics =
